wheezy: test PURE installation without gems

+++ templates/Debian-7.0-rc1-amd64-netboot/ruby.sh
@@ -1,10 +1,18 @@
+if [ -f .veewee_params ]
+then
+  . .veewee_params
+fi
 # Install Ruby from packages
 apt-get -y install ruby ruby-dev libopenssl-ruby1.8 irb ri rdoc
 
-# Install Rubygems from source
-rg_ver=1.8.22
-curl -o /tmp/rubygems-${rg_ver}.zip \
-  "http://production.cf.rubygems.org/rubygems/rubygems-${rg_ver}.zip"
-(cd /tmp && unzip rubygems-${rg_ver}.zip && \
-  cd rubygems-${rg_ver} && ruby setup.rb --no-format-executable)
-rm -rf /tmp/rubygems-${rg_ver} /tmp/rubygems-${rg_ver}.zip
+if [ "$PURE" == "yes" ]; then
+  apt-get -y install rubygems
+else
+  # Install Rubygems from source
+  rg_ver=1.8.22
+  curl -o /tmp/rubygems-${rg_ver}.zip \
+    "http://production.cf.rubygems.org/rubygems/rubygems-${rg_ver}.zip"
+  (cd /tmp && unzip rubygems-${rg_ver}.zip && \
+    cd rubygems-${rg_ver} && ruby setup.rb --no-format-executable)
+  rm -rf /tmp/rubygems-${rg_ver} /tmp/rubygems-${rg_ver}.zip
+fi
